requirements.txt
mcp
spotipy
python-dotenv
googlemaps
